requirements.txt
python-dotenv>=1.0.0
feedparser>=6.0.11
playwright>=1.46
